requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
async-timeout==5.0.1
    # via redis
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
coloredlogs==15.0.1
    # via onnxruntime
exceptiongroup==1.3.0
    # via anyio
fastembed==0.7.3
    # via -r requirements.in
filelock==3.19.1
    # via huggingface-hub
flatbuffers==25.9.23
    # via onnxruntime
fsspec==2025.9.0
    # via huggingface-hub
grpcio==1.75.1
    # via qdrant-client
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hf-xet==1.1.10
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via qdrant-client
huggingface-hub==0.35.3
    # via
    #   fastembed
    #   tokenizers
humanfriendly==10.0
    # via coloredlogs
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
kafka-python==2.2.15
    # via -r requirements.in
loguru==0.7.3
    # via fastembed
mmh3==5.2.0
    # via fastembed
mpmath==1.3.0
    # via sympy
numpy==2.2.6
    # via
    #   fastembed
    #   onnxruntime
    #   qdrant-client
onnxruntime==1.23.0
    # via
    #   -r requirements.in
    #   fastembed
packaging==25.0
    # via
    #   huggingface-hub
    #   onnxruntime
pillow==11.3.0
    # via fastembed
portalocker==3.2.0
    # via qdrant-client
protobuf==6.32.1
    # via
    #   onnxruntime
    #   qdrant-client
py-rust-stemmers==0.1.5
    # via fastembed
pydantic==2.11.9
    # via qdrant-client
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.3
    # via huggingface-hub
qdrant-client==1.15.1
    # via -r requirements.in
redis==6.4.0
    # via -r requirements.in
requests==2.32.5
    # via
    #   fastembed
    #   huggingface-hub
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sympy==1.14.0
    # via onnxruntime
tokenizers==0.22.1
    # via fastembed
tqdm==4.67.1
    # via
    #   fastembed
    #   huggingface-hub
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   grpcio
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
